% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{and}
\alias{and}
\alias{not}
\alias{or}
\title{And}
\usage{
and(v1, v2)

or(v1, v2)

not(v)
}
\arguments{
\item{v1}{A vector of fuzzy set scores of cases}

\item{v2}{A vector of fuzzy set scores of cases}

\item{v}{A vector of fuzzy set scores of cases}
}
\value{
the fuzzy set scores of the logical conjunction of v1 and v2 for each case, i.e.
the minimum in each component

the fuzzy set scores of the logical disjunction of v1 and v2 for each case, i.e.
the maximum in each component

the fuzzy set scores of the negation of v for each case, i.e. 1-v
}
\description{
Logical 'and' of two conditions

Logical 'or' of two conditions

Logical 'not' of a condition
}
\examples{
and(c(0,0.5,1), c(0.25, 0.75, 0.75))
or(c(0,0.5,1), c(0.25, 0.75, 0.75))
not(c(0,0.5,1))

}

