% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{format_dnf}
\alias{format_dnf}
\title{Rewrite a list of clauses to a string containing a Boolean formula in disjunctive normal form}
\usage{
format_dnf(dnf)
}
\arguments{
\item{dnf}{list of clauses}
}
\value{
string containing a Boolean formula in disjunctive normal form
}
\description{
Rewrite a list of clauses to a string containing a Boolean formula in disjunctive normal form
}

