% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xyplot.R
\name{xyplot}
\alias{xyplot}
\title{Plot fuzzy set score of two sets against each other}
\usage{
xyplot(x, y, data, labels = FALSE, main.diagonal = TRUE,
  anti.diagonal = FALSE)
}
\arguments{
\item{x}{Formula that describes the fuzzy set to plot along the x axis}

\item{y}{Formula that describes the fuzzy set to plot along the y axis}

\item{data}{Data set of basic fuzzy set scores}

\item{labels}{flag whether to label individual points with the case names}

\item{main.diagonal}{flag whether to plot the main diagonal}

\item{anti.diagonal}{flag whether to plot the anti diagonal}
}
\value{
the \code{ggplot} plot object
}
\description{
Plot fuzzy set score of two sets against each other
}
\examples{
require(QCA)
data(d.urban)
xyplot("MLC", "WSR", d.urban)
}

